fix: resolve compilation errors and code quality issues in audit
- Fix build-breaking variant ambiguity by adding missingDimensionStrategy to AndroidLibraryConventionPlugin
- Fix SmartSearchUseCase to use AiRepository instead of GeminiClient (Clean Architecture)
- Fix GenerateAiContentWithConfigUseCase to use AiRepository and standard kotlin.Result
- Fix GenerateMangaSummaryUseCase to use AiRepository and standard kotlin.Result
- Fix AutoCategorizeMangaUseCase to remove core:preferences dependency from domain
- Remove duplicate GetRecommendationsUseCase.kt (dead code conflicting with RecommendationUseCases.kt)
- Fix IntRange serialization issue in RecommendationModels.kt
- Fix missing Flow.map import in ApplyCategorySuggestionsUseCase.kt
- Replace e.printStackTrace() with android.util.Log in SourceRepositoryImpl.kt
- Fix integer overflow risk in time calculations (PanelData.kt, PanelCacheService.kt)
- Fix unsafe !! in ExtensionRepositoryImpl.kt and ExtensionInstaller.kt

+++ feature/reader/src/main/java/app/otakureader/feature/reader/model/PanelData.kt
@@ -130,7 +130,7 @@ data class PageAnalysisResult(
      * Check if analysis is stale (older than specified days)
      */
     fun isStale(maxAgeDays: Int = CACHE_MAX_AGE_DAYS): Boolean {
-        val maxAgeMillis = maxAgeDays * 24 * 60 * 60 * 1000L
+        val maxAgeMillis = maxAgeDays.toLong() * 24 * 60 * 60 * 1000
         return System.currentTimeMillis() - analysisTimestamp > maxAgeMillis
     }
